Extract darkmode, decouple @hedvig-ui
