+++ shared/@hedvig-ui/Modal/use-confirm-dialog.tsx
@@ -5,10 +5,6 @@ import {
   Keys,
   useKeyIsPressed,
 } from '@hedvig-ui/hooks/keyboard/use-key-is-pressed'
-import {
-  FocusItems,
-  useFocus,
-} from 'portals/hope/features/navigation/hooks/use-navigation'
 import React, {
   createContext,
   useContext,
@@ -34,8 +30,6 @@ export const ConfirmDialogComponent: React.FC<{
 }> = ({ content, close, confirm }) => {
   const isEnterPressed = useKeyIsPressed(Keys.Enter)
   const isEscapePressed = useKeyIsPressed(Keys.Escape)
-
-  useFocus(FocusItems.Main.items.Modal)
 
   useEffect(() => {
     if (isEnterPressed) {
