--- shared/@hedvig-ui/hooks/use-darkmode.ts
@@ -1,13 +1,0 @@
-import { createContext, useContext } from 'react'
-
-export const getDefaultIsDarkmode = () =>
-  window.localStorage.getItem('hedvig:theming:darkmode') === 'true'
-
-export const UseDarkmode = createContext({
-  isDarkmode: getDefaultIsDarkmode(),
-  setIsDarkmode: (_isDarkmode: boolean) => {
-    /* noop */
-  },
-})
-
-export const useDarkmode = () => useContext(UseDarkmode)
